Removed fetching SanitySettings to providers
